add FCProject config file to speed up global vars import
See https://github.com/nm2107/FreeCADMacroImportGlobalVars
This file is used with a macro from the above link to speed up the
import of the global vars into the active document.

As the project is composed of many linked documents, each time we create
a new document, we have to import the global vars. By using this macro,
the import process is now faster.